// src/types/componentTypes/adminSettings.ts
export interface WebsiteSettingsFormValues {
  companyName: string;
  logo: FileList;
}

export interface FooterSettingsFormValues {
  address: string;
  location: string;
  phoneIndia: string;
  phoneUAE: string;
  facebookUrl: string;
  xUrl: string;
  instagramUrl: string;
  companyEmail: string;
  timingInida: string;
  timingUae: string;
}

export interface AboutSettingsFormValues {
  aboutText: string;
}